refactor PaginationConcern
