+++ client/app/inbox/pages/InboxPage.jsx
@@ -137,7 +137,7 @@ class InboxMessagesPage extends React.PureComponent {
       <Pagination
         currentPage={this.props.pagination.current_page}
         currentCases={rowObjects.length}
-        totalCases={this.props.pagination.total_messages}
+        totalCases={this.props.pagination.total_items}
         totalPages={this.props.pagination.total_pages}
         pageSize={this.props.pagination.page_size}
         updatePage={pageUpdater} />
